combine question reproductions into one spec
