--- e2e/test/scenarios/question/reproductions/25144-saved-questions-first-question.cy.spec.js
@@ -1,60 +1,0 @@
-import { modal, popover, restore, describeOSS } from "e2e/support/helpers";
-
-// this is only testable in OSS because EE always has models from auditv2
-describeOSS("issue 25144", { tags: "@OSS" }, () => {
-  beforeEach(() => {
-    restore("setup");
-    cy.signInAsAdmin();
-    cy.intercept("POST", "/api/card").as("createCard");
-    cy.intercept("PUT", "/api/card/*").as("updateCard");
-  });
-
-  it("should show Saved Questions section after creating the first question (metabase#25144)", () => {
-    cy.visit("/");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("New").click();
-    popover().findByText("Question").click();
-    popover().findByText("Orders").click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Save").click();
-    modal().findByLabelText("Name").clear().type("Orders question");
-    modal().button("Save").click();
-    cy.wait("@createCard");
-    cy.wait(100);
-    modal().button("Not now").click();
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("New").click();
-    popover().findByText("Question").click();
-    popover().findByText("Saved Questions").click();
-    popover().findByText("Orders question").should("be.visible");
-  });
-
-  it("should show Models section after creation the first model (metabase#24878)", () => {
-    cy.visit("/");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("New").click();
-    popover().findByText("Question").click();
-    popover().findByText("Orders").click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Save").click();
-    modal().findByLabelText("Name").clear().type("Orders model");
-    modal().button("Save").click();
-    cy.wait("@createCard");
-    cy.wait(100);
-    modal().button("Not now").click();
-
-    cy.findByLabelText("Move, archive, and more...").click();
-    popover().findByText("Turn into a model").click();
-    modal().button("Turn this into a model").click();
-    cy.wait("@updateCard");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("New").click();
-    popover().findByText("Question").click();
-    popover().findByText("Models").click();
-    popover().findByText("Orders model").should("be.visible");
-  });
-});
