--- e2e/test/scenarios/question/reproductions/28221-missing-custom-field-metadata.cy.spec.js
@@ -1,48 +1,0 @@
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { restore } from "e2e/support/helpers";
-
-const { PRODUCTS_ID, PRODUCTS, ORDERS_ID, ORDERS } = SAMPLE_DATABASE;
-
-describe("issue 28221", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-  });
-
-  it("should be able to select see notebook view even if a question custom field metadata is missing#27462", () => {
-    const questionName = "Reproduce 28221";
-    const customFieldName = "Non-existing field";
-    const questionDetails = {
-      name: questionName,
-      query: {
-        "source-table": ORDERS_ID,
-        joins: [
-          {
-            fields: "all",
-            "source-table": PRODUCTS_ID,
-            condition: [
-              "=",
-              ["field", ORDERS.PRODUCT_ID, null],
-              ["field", PRODUCTS.ID, { "join-alias": "Products" }],
-            ],
-            alias: "Products",
-          },
-        ],
-        expressions: {
-          [customFieldName]: ["field", 9999, null],
-        },
-      },
-    };
-
-    cy.createQuestion(questionDetails).then(({ body }) => {
-      const questionId = body.id;
-
-      cy.visit(`/question/${questionId}/notebook`);
-    });
-
-    cy.findByDisplayValue(questionName).should("be.visible");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText(customFieldName).should("be.visible");
-  });
-});
