--- e2e/test/scenarios/question/reproductions/28874-notebook-pivot.cy.spec.js
@@ -1,39 +1,0 @@
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { restore, visitQuestionAdhoc } from "e2e/support/helpers";
-
-const { ORDERS, ORDERS_ID } = SAMPLE_DATABASE;
-
-const questionDetails = {
-  name: "28874",
-  display: "pivot",
-  dataset_query: {
-    database: SAMPLE_DB_ID,
-    type: "query",
-    query: {
-      "source-table": ORDERS_ID,
-      aggregation: [["count"]],
-      breakout: [
-        ["field", ORDERS.CREATED_AT, { "temporal-unit": "year" }],
-        ["field", ORDERS.PRODUCT_ID, null],
-      ],
-    },
-  },
-};
-
-describe("issue 28874", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsNormalUser();
-  });
-
-  it("should allow to modify a pivot question in the notebook (metabase#28874)", () => {
-    visitQuestionAdhoc(questionDetails, { mode: "notebook" });
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Product ID").parent().icon("close").click();
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Product ID").should("not.exist");
-  });
-});
