--- e2e/test/scenarios/question/reproductions/38176-revert-to-previous-does-not-honor-parameters.cy.spec.js
@@ -1,51 +1,0 @@
-import { restore, questionInfoButton, rightSidebar } from "e2e/support/helpers";
-
-describe("issue 38176", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsNormalUser();
-    cy.intercept("PUT", "/api/card/**").as("updateQuestion");
-  });
-
-  it("restoring a question to a previous version should preserve the variables (metabase#38176)", () => {
-    cy.createNativeQuestion(
-      {
-        name: "38176",
-        native: {
-          query:
-            'SELECT "COUNTRY" from "ACCOUNTS" WHERE country = {{ country }} LIMIT 5',
-          "template-tags": {
-            country: {
-              type: "text",
-              id: "dd06cd10-596b-41d0-9d6e-94e98ceaf989",
-              name: "country",
-              "display-name": "Country",
-            },
-          },
-        },
-      },
-      { visitQuestion: true },
-    );
-
-    cy.findByPlaceholderText("Country").type("NL");
-
-    cy.findByTestId("query-builder-main").button("Get Answer").click();
-
-    questionInfoButton().click();
-    rightSidebar().within(() => {
-      cy.findByText("History");
-
-      cy.findByPlaceholderText("Add description")
-        .type("This is a question")
-        .blur();
-
-      cy.wait("@updateQuestion");
-      cy.findByText(/added a description/i);
-      cy.findByTestId("question-revert-button").click();
-
-      cy.findByText(/reverted to an earlier version/i).should("be.visible");
-    });
-
-    cy.findAllByRole("gridcell").should("contain", "NL");
-  });
-});
